+++ integrations/argocd/CHANGELOG.md
@@ -9,10 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Port_Ocean 0.1.35 (2024-04-04)
 
+### Deprecations
+- Added deprecation warnings to the deployment-history and kubernetes-resource kind, urging users to utilize the itemsToParse functionality instead
+
 ### Improvements
 
 - Reveresed the relation between cluster and namespace, and other general enhancements on blueprints (PORT-7550)
-- Updated the default mapping for deployment-history and kubernetes-resource to use the itemsToParse functionality and introduced a deprecation warning for these two kinds
+- Updated the default mapping for deployment-history and kubernetes-resource kind to reuse the application kind's response and parse items using the itemsToParse functionality for improved efficiency
 
 
 # Port_Ocean 0.1.34 (2024-04-01)